Disrespect passkey discoverability
According to the WebAuthn spec, there is no mechanism to enforce the
creation of non-discoverable credentials. Therefore it is the
responsiblity of relying parties to properly handle a passkey created
with residentKey=discouraged being used in an usernameless login flow.

Additionally, the current behavior is likely to cause pain and confusion
for end users who expects to be able to use such passkeys for
usernameless login.

Fix this by disrespecting the discoverability of existing passkeys and
always creating new passkeys as discoverable.